Move clients to api_v2.clients
